# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Third Space Risk OS",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/ingest/{venue_id}/pos": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Ingest Pos",
        "operationId": "ingest_pos_api_v1_ingest__venue_id__pos_post",
        "parameters": [
          {
            "name": "venue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/POSEvent"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingest/{venue_id}/camera": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Ingest Camera",
        "operationId": "ingest_camera_api_v1_ingest__venue_id__camera_post",
        "parameters": [
          {
            "name": "venue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CameraEvent"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingest/{venue_id}/staffing": {
      "post": {
        "tags": [
          "ingestion"
        ],
        "summary": "Ingest Staffing",
        "operationId": "ingest_staffing_api_v1_ingest__venue_id__staffing_post",
        "parameters": [
          {
            "name": "venue_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Venue Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffingEvent"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CameraEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id",
            "description": "Unique identifier for the event"
          },
          "venue_id": {
            "type": "string",
            "title": "Venue Id",
            "description": "The unique ID of the venue (e.g., elsewhere-brooklyn)"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType",
            "default": "camera"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "When the event occurred at the source"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At",
            "description": "When the event hit our ingestion API"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Arbitrary source metadata (firmware, sensor IDs, etc.)"
          },
          "payload": {
            "$ref": "#/components/schemas/CameraEventPayload"
          }
        },
        "type": "object",
        "required": [
          "venue_id",
          "payload"
        ],
        "title": "CameraEvent"
      },
      "CameraEventPayload": {
        "properties": {
          "zone_id": {
            "type": "string",
            "title": "Zone Id"
          },
          "person_count": {
            "type": "integer",
            "title": "Person Count"
          },
          "detections": {
            "items": {
              "$ref": "#/components/schemas/Detection"
            },
            "type": "array",
            "title": "Detections"
          },
          "aggression_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Aggression Score",
            "default": 0.0
          },
          "anomaly_detected": {
            "type": "boolean",
            "title": "Anomaly Detected",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "zone_id",
          "person_count",
          "detections"
        ],
        "title": "CameraEventPayload"
      },
...
      "Detection": {
        "properties": {
          "label": {
            "type": "string",
            "title": "Label"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "box": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Box"
          }
        },
        "type": "object",
        "required": [
          "label",
          "confidence",
          "box"
        ],
        "title": "Detection"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "POSEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id",
            "description": "Unique identifier for the event"
          },
          "venue_id": {
            "type": "string",
            "title": "Venue Id",
            "description": "The unique ID of the venue (e.g., elsewhere-brooklyn)"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType",
            "default": "pos"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "When the event occurred at the source"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At",
            "description": "When the event hit our ingestion API"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Arbitrary source metadata (firmware, sensor IDs, etc.)"
          },
          "payload": {
            "$ref": "#/components/schemas/POSEventPayload"
          }
        },
        "type": "object",
        "required": [
          "venue_id",
          "payload"
        ],
        "title": "POSEvent"
      },
      "POSEventPayload": {
        "properties": {
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "total_amount": {
            "type": "number",
            "title": "Total Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "USD"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/POSItem"
            },
            "type": "array",
            "title": "Items"
          },
          "payment_method": {
            "type": "string",
            "title": "Payment Method"
          },
          "staff_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Staff Id"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "total_amount",
          "items",
          "payment_method"
        ],
        "title": "POSEventPayload"
      },
      "POSItem": {
        "properties": {
          "sku": {
            "type": "string",
            "title": "Sku"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "quantity": {
            "type": "integer",
            "title": "Quantity"
          },
          "price_total": {
            "type": "number",
            "title": "Price Total"
          },
          "category": {
            "type": "string",
            "title": "Category"
          }
        },
        "type": "object",
        "required": [
          "sku",
          "name",
          "quantity",
          "price_total",
          "category"
        ],
        "title": "POSItem"
      },
...
      "SourceType": {
        "type": "string",
        "enum": [
          "pos",
          "camera",
          "staffing",
          "iot"
        ],
        "title": "SourceType"
      },
      "StaffingEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "format": "uuid",
            "title": "Event Id",
            "description": "Unique identifier for the event"
          },
          "venue_id": {
            "type": "string",
            "title": "Venue Id",
            "description": "The unique ID of the venue (e.g., elsewhere-brooklyn)"
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType",
            "default": "staffing"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "When the event occurred at the source"
          },
          "received_at": {
            "type": "string",
            "format": "date-time",
            "title": "Received At",
            "description": "When the event hit our ingestion API"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Arbitrary source metadata (firmware, sensor IDs, etc.)"
          },
          "payload": {
            "$ref": "#/components/schemas/StaffingEventPayload"
          }
        },
        "type": "object",
        "required": [
          "venue_id",
          "payload"
        ],
        "title": "StaffingEvent"
      },
      "StaffingEventPayload": {
        "properties": {
          "staff_id": {
            "type": "string",
            "title": "Staff Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "location_zone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Zone"
          }
        },
        "type": "object",
        "required": [
          "staff_id",
          "name",
          "role",
          "action"
        ],
        "title": "StaffingEventPayload"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}